Backend/mcp-server/src/tools/vertex_ai_enhance.py
"""
CarFin MCP Tool: Vertex AI Enhancement
Vertex AI 서비스 통합으로 추천 시스템 고도화
"""

import asyncio
import logging
from typing import Dict, List, Any, Optional
from google.cloud import aiplatform
from google.cloud.aiplatform.gapic.schema import predict
import numpy as np

logger = logging.getLogger("CarFin-MCP.VertexAI")

class VertexAIEnhancer:
    """Vertex AI 서비스 통합 클래스"""

    def __init__(self, project_id: str, region: str = "asia-northeast1"):
        self.project_id = project_id
        self.region = region
        aiplatform.init(project=project_id, location=region)

    async def get_text_embedding(self, text: str) -> List[float]:
        """텍스트 임베딩 생성"""
        try:
            # Vertex AI Text Embeddings API 사용
            endpoint = aiplatform.Endpoint(
                endpoint_name=f"projects/{self.project_id}/locations/{self.region}/endpoints/text-embedding"
            )

            instances = [{"content": text}]
            response = await endpoint.predict(instances=instances)

            return response.predictions[0]["embeddings"]["values"]

        except Exception as e:
            logger.error(f"❌ Vertex AI 임베딩 생성 실패: {e}")
            return []

    async def enhance_user_profile(self, user_data: Dict[str, Any]) -> Dict[str, Any]:
        """사용자 프로필 Vertex AI 임베딩으로 강화"""
        try:
            # 사용자 설명 텍스트 생성
            user_description = f"""
            나이: {user_data.get('age', 30)}세
            직업: {user_data.get('occupation', '직장인')}
            예산: {user_data.get('budget', 3000)}만원
            선호 브랜드: {user_data.get('preferred_brands', ['현대', '기아'])}
            용도: {user_data.get('usage', '출퇴근')}
            중요 요소: {user_data.get('priorities', ['연비', '안전성'])}
            """

            # Vertex AI 임베딩 생성
            user_embedding = await self.get_text_embedding(user_description)

            # 기존 프로필에 임베딩 추가
            enhanced_profile = user_data.copy()
            enhanced_profile["vertex_ai_embedding"] = user_embedding
            enhanced_profile["embedding_dimension"] = len(user_embedding)

            return enhanced_profile

        except Exception as e:
            logger.error(f"❌ 사용자 프로필 강화 실패: {e}")
            return user_data

    async def enhance_vehicle_data(self, vehicle_data: Dict[str, Any]) -> Dict[str, Any]:
        """차량 데이터 Vertex AI 임베딩으로 강화"""
        try:
            # 차량 설명 텍스트 생성
            vehicle_description = f"""
            브랜드: {vehicle_data.get('brand', '')}
            모델: {vehicle_data.get('model', '')}
            연식: {vehicle_data.get('year', 2020)}년
            주행거리: {vehicle_data.get('mileage', 0)}km
            연료: {vehicle_data.get('fuel_type', '가솔린')}
            변속기: {vehicle_data.get('transmission', '자동')}
            가격: {vehicle_data.get('price', 0)}만원
            상태: {vehicle_data.get('condition', '양호')}
            """

            # Vertex AI 임베딩 생성
            vehicle_embedding = await self.get_text_embedding(vehicle_description)

            # 기존 데이터에 임베딩 추가
            enhanced_vehicle = vehicle_data.copy()
            enhanced_vehicle["vertex_ai_embedding"] = vehicle_embedding
            enhanced_vehicle["embedding_dimension"] = len(vehicle_embedding)

            return enhanced_vehicle

        except Exception as e:
            logger.error(f"❌ 차량 데이터 강화 실패: {e}")
            return vehicle_data

    async def calculate_semantic_similarity(
        self,
        user_embedding: List[float],
        vehicle_embedding: List[float]
    ) -> float:
        """시맨틱 유사도 계산"""
        try:
            if not user_embedding or not vehicle_embedding:
                return 0.0

            # 코사인 유사도 계산
            user_array = np.array(user_embedding)
            vehicle_array = np.array(vehicle_embedding)

            dot_product = np.dot(user_array, vehicle_array)
            norm_user = np.linalg.norm(user_array)
            norm_vehicle = np.linalg.norm(vehicle_array)

            if norm_user == 0 or norm_vehicle == 0:
                return 0.0

            similarity = dot_product / (norm_user * norm_vehicle)
            return float(similarity)

        except Exception as e:
            logger.error(f"❌ 시맨틱 유사도 계산 실패: {e}")
            return 0.0

    async def gemini_pro_analysis(self, context: str, question: str) -> str:
        """Gemini Pro를 통한 고급 분석"""
        try:
            # Vertex AI Gemini Pro 모델 사용
            model = aiplatform.gapic.PredictionServiceClient()

            prompt = f"""
            컨텍스트: {context}

            질문: {question}

            중고차 전문가로서 데이터를 분석하고 구체적인 인사이트를 제공해주세요.
            """

            # Gemini Pro 추론 (실제 구현시 Vertex AI SDK 사용)
            response = await self._call_gemini_pro(prompt)
            return response

        except Exception as e:
            logger.error(f"❌ Gemini Pro 분석 실패: {e}")
            return "분석을 수행할 수 없습니다."

    async def _call_gemini_pro(self, prompt: str) -> str:
        """Gemini Pro API 호출 (모킹)"""
        # 실제 구현시 Vertex AI Gemini Pro API 사용
        await asyncio.sleep(0.5)  # API 호출 시뮬레이션
        return f"Gemini Pro 분석 결과: {prompt[:100]}..."

# MCP Tool 함수
async def vertex_ai_enhance_tool(params: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
    """Vertex AI 강화 MCP Tool"""
    try:
        enhancer = VertexAIEnhancer(
            project_id=context.get("gcp_project_id", "carfin-ai-production"),
            region="asia-northeast1"
        )

        task_type = params.get("task_type", "enhance_recommendation")

        if task_type == "enhance_user_profile":
            user_data = params.get("user_data", {})
            result = await enhancer.enhance_user_profile(user_data)

        elif task_type == "enhance_vehicle_data":
            vehicle_data = params.get("vehicle_data", {})
            result = await enhancer.enhance_vehicle_data(vehicle_data)

        elif task_type == "semantic_similarity":
            user_emb = params.get("user_embedding", [])
            vehicle_emb = params.get("vehicle_embedding", [])
            similarity = await enhancer.calculate_semantic_similarity(user_emb, vehicle_emb)
            result = {"semantic_similarity": similarity}

        elif task_type == "gemini_analysis":
            context_text = params.get("context", "")
            question = params.get("question", "")
            analysis = await enhancer.gemini_pro_analysis(context_text, question)
            result = {"gemini_analysis": analysis}

        else:
            result = {"error": f"Unknown task_type: {task_type}"}

        return {
            "success": True,
            "result": result,
            "vertex_ai_enhanced": True
        }

    except Exception as e:
        logger.error(f"❌ Vertex AI 강화 도구 실패: {e}")
        return {
            "success": False,
            "error": str(e),
            "vertex_ai_enhanced": False
        }