Backend/mcp-server/requirements.txt
# CarFin-MCP Server Dependencies - NCF 포함 버전

# 웹 프레임워크 (필수)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 데이터베이스 (필수)
asyncpg==0.29.0

# AI/ML (NCF 모델용)
torch==2.1.1
numpy==1.24.3
scikit-learn==1.3.2

# Google Cloud AI/ML
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0

# HTTP 클라이언트 (필수)
httpx==0.25.2

# 유틸리티 (필수)
python-dotenv==1.0.0
python-multipart==0.0.6